requirements.txt
flake8==6.0.0
hydra-core==1.3.2
imbalanced-learn==0.10.1
ipykernel==6.21.3
ipython==8.12.0
jupyter-core==5.3.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.2
omegaconf==2.3.0
pandas==1.5.3
pip==20.2.4
plotly==5.13.1
pre-commit==3.1.1
python-dotenv==1.0.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
wandb==0.14.0
xgboost==1.7.4
# git+https://github.com/mklasby/ingeniator.git 
